+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix scrolling containers while dragging in Sortable component
 - Fix initializing components with jQuery elements
 - Fix show/hide behaviour in Drop component 
+- Fix Accordion component will scroll title into view if needed
 
 ## 3.2.2 (October 23, 2019)